cleaned up the project and restructured files
