+++ .expo/packager-info.json
@@ -2,9 +2,9 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 15284,
+  "packagerPid": 7817,
   "expoServerNgrokUrl": "https://c3-8rp.anonymous.snapchef.exp.direct",
   "packagerNgrokUrl": "https://packager.c3-8rp.anonymous.snapchef.exp.direct",
-  "ngrokPid": 14988,
+  "ngrokPid": 7843,
   "webpackServerPort": null
 }
