+++ .expo/settings.json
@@ -1,7 +1,7 @@
 {
   "dev": true,
   "minify": false,
-  "https": false,
+  "https": true,
   "hostType": "lan",
   "lanType": "ip",
   "urlRandomness": "c3-8rp"
