+++ .gitignore
@@ -1,1 +1,2 @@
 node_modules
+config.js
