--- helpers/config.js
@@ -1,1 +1,0 @@
-
