+++ pages/CameraView.js
