+++ pages/recipeListPage.js
@@ -1,23 +1,8 @@
 import React, { useState, useEffect } from "react";
-import {
-  StyleSheet,
-  Text,
-  View,
-  Image,
-  ScrollView,
-  ImageBackground,
-  TouchableHighlight,
-  Linking,
-} from "react-native";
-import {
-  TextInput,
-  List,
-  Button,
-  Card,
-  Title,
-  Paragraph,
-} from "react-native-paper";
-import Recipe from "./components/Recipe";
+import { StyleSheet, View, ScrollView, Linking } from "react-native";
+import { TextInput, Button } from "react-native-paper";
+import Recipe from "../components/Recipe";
+import config from "../config";
 const styles = StyleSheet.create({
   container: {
     flex: 1,
@@ -39,21 +24,13 @@ const styles = StyleSheet.create({
 });
 
 function recipeListPage({ navigation, route }) {
-  const {image} = route.params;
+  const { image } = route.params;
 
   const [recipes, setRecipes] = useState([]);
   const googleInfo = "";
   const [searchBar, setSearchBar] = useState(googleInfo);
-  const config = {
-    recipe: {
-      appID: "42589782",
-      key: "ac119e6e01c62306b308b4232bf9403a",
-    },
-  };
 
-  async function callGoogleAPI()
-  {
-
+  async function callGoogleAPI() {
     let body = JSON.stringify({
       requests: [
         {
@@ -67,40 +44,35 @@ function recipeListPage({ navigation, route }) {
             { type: "SAFE_SEARCH_DETECTION", maxResults: 5 },
             { type: "IMAGE_PROPERTIES", maxResults: 5 },
             { type: "CROP_HINTS", maxResults: 5 },
-            { type: "WEB_DETECTION", maxResults: 5 }
+            { type: "WEB_DETECTION", maxResults: 5 },
           ],
           image: {
-            content: image
-              
-          }
-        }
-      ]
-    }); 
+            content: image,
+          },
+        },
+      ],
+    });
 
-    const response = await fetch( 
-      "https://vision.googleapis.com/v1/images:annotate?key=XXXXX",
+    const response = await fetch(
+      `https://vision.googleapis.com/v1/images:annotate?key=${config.googleAPI.key}`,
       {
         headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json"
+          Accept: "application/json",
+          "Content-Type": "application/json",
         },
         method: "POST",
-        body: body
+        body: body,
       }
     );
-      
+
     const results = await response.json();
 
     console.log(results);
-
-    
   }
-
-
 
   async function getRecipes() {
     await fetch(
-      `https://api.edamam.com/search?q=${searchBar}&app_id=${config.recipe.appID}&app_key=${config.recipe.key}`
+      `https://api.edamam.com/search?q=${searchBar}&app_id=${config.edamamAPI.appID}&app_key=${config.edamamAPI.key}`
     )
       .then((response) => {
         response.json().then((data) => setRecipes(data.hits));
